refactor: change package structure to be more clear and less verbose
