+++ src/main/kotlin/com/github/frozensync/database/FirestoreDocument.kt
@@ -1,4 +1,4 @@
-package com.github.frozensync.persistence.firestore
+package com.github.frozensync.database
 
 /**
  * Enables the no-arg compiler plugin to provide a no-arg constructor to annotated classes.
