--- src/main/kotlin/com/github/frozensync/raspberrypi/Module.kt
@@ -1,7 +1,0 @@
-package com.github.frozensync.raspberrypi
-
-import org.koin.dsl.module
-
-val raspberryPiModule = module {
-    single<RaspberryPiService> { RaspberryPiServiceImpl(get(), get()) }
-}
