+++ src/main/kotlin/com/github/frozensync/tournament/TournamentModule.kt
@@ -1,9 +1,13 @@
 package com.github.frozensync.tournament
 
+import com.github.frozensync.tournament.raspberrypi.RaspberryPiService
+import com.github.frozensync.tournament.raspberrypi.RaspberryPiServiceImpl
 import org.koin.dsl.module
 
 val tournamentModule = module {
     single<TournamentService> { TournamentServiceImpl(get(), get()) }
+
+    single<RaspberryPiService> { RaspberryPiServiceImpl(get(), get()) }
 
     single { ScorerService(get()) }
     single { ScorerServer(get(), get()) }
