+++ src/main/kotlin/com/github/frozensync/tournament/raspberrypi/RaspberryPi.kt
@@ -1,4 +1,4 @@
-package com.github.frozensync.raspberrypi
+package com.github.frozensync.tournament.raspberrypi
 
 import com.github.frozensync.DeviceId
